leds: pca963x: Use the proper bits for the invert flag
This patch applies the proper bit for the invert flag. Due to an
oversight we where off by 1 bit after the totem_pole flag which caused
us to set the wrong output when having an inverted flag. This resulted
in the output to be the inverse of the inverted.

Additionally some mutex locking improvements.

EM-132

Signed-off-by: Olliver Schinagl <o.schinagl@ultimaker.com>