leds: pca963x: Remove global dim control (gdc)
The led driver has two features to keep leds uniform. Global frequency
to blink leds in sync via hardware and global dim control, to brighten
and dim leds in sync. The later was being controlled in the set_blink
routine, which does not make a lot of sense, as we do not want to change
brightness at all, just blink. The calculation used to setup the global
dim values was based on the period, e.g. faster blinking equals brighter
output.

This patch removes gdc altogether since currently it does not fit in the
core led class. There is no linking of leds (the blink is already quirky
as is) and no linked dimming is supported at all (no knob for this
even).

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>