regulator: axp20x: AXP209: Add support for set_ramp_delay
The AXP209 supports ramping up voltages on several regulators such as
DCDC2 and LDO3.

This patch adds preliminary support for these 2 regulators. Note that
the voltage ramp only works when the regulator is already enabled. E.g.
when going from say 0.7 V to 3.6 V. When turning on the regulator, no
voltage ramp is performed by the hardware.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>