regulator: axp20x: add software based soft_start for LDO3
In the past, there have been words that if LDO3 is disabled in u-boot,
but enabled in the DTS, the axp driver would fail to continue/hang.

What really happened however, was that the AXP shuts down, without
notice, without setting an interrupt. This is caused when LDO3 gets
overloaded, for example with large capacitors on the LDO3 output.

Normally, we would expect that the AXP would source 200 mA as per
datasheet and set and trigger an interrupt.

As a work-around, we use the soft-start constraint of the regulator node
to first bring up the LDO3 to the lowest possible voltage and then
enable the LDO. After that, we can set the requested voltage as per
normal.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>