pwm: core: do not block apply->state on period
The old pwm_config call, did some parameter sanitation such as
preventing to change the pwm config if the period is 0.

With the new atomic framework, we now not only apply configuration
parameters, but also the polarity setting.

The polarity setting however can be applied both with an enabled or
disabled PWM, in other words, setting the polarity is perfectly
acceptable when the period is 0.

This patch removes the check if the period is valid and instead, sets
the state disabled if the period is 0. It now thus becomes possible to
set the polarity, before changing any other PWM parameters.

The reason this is needed is that if someone calls pwm_set_polarity
with the period set to 0, pwm_set_polarity would fail. Having period of 0
is however perfectly acceptable here.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>