leds: pca963x: Do not load driver if chip not found
Currently, the driver always successfully manages to load,
even if there is no actual chip present (or communication
is not possible over i2c).

This has the side-effect, that the node is always available in
sysfs, and writing data to it, results in i2c errors.

To improve behavior here, we try to read the first byte from
the device, and if this was okay, we atleast know there is a
device on the bus.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>