leds: pca963x: abort probe if device is not connected
Currently, the driver always successfully manages to load,
even if there is no actual chip present (or communication
is not possible over i2c).

This has the side-effect, that the node is always available in
sysfs, and writing data to it which results in i2c errors.

To improve behavior here, we try to read the first byte from
the device, and if this was okay, we at least know there is some
device on the bus, even though we do not know which one.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>