ASoC: sun4i-is: also check for NULL on reset pin request
Commit 2ad6f30de7087 ("ASoC: sun4i-i2s: Add quirks to handle a31 compatible")
added support for reset control but support for sun4i was overlooked. For
these older SoC's it is quite common not have this reset control.

The reset control is first requested, and later checked if it IS_ERR.
If however there is no reset control configured, we can get NULL and
thus the checks pass with a NULL pointer passed the functions
afterwards.

Lets use IS_ERR_OR_NULL to ensure the reset control is only
(de)asserted when we actually have a valid reset control.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>