[WIP-do-not-merge] drm/sunxi: Split off hdmi-i2c as a separated driver
With this patch, the hdmi-i2c peripheral is split off into its own
identity. By doing so, this gives board designers and users various new
uses.

It is now possible to use the ddc-i2c-bus property on the
display node to actually choose an I2C controller. E.g. if a board
designer chooses to use a different I2C controller for the EDID eeprom,
they are free to do so.

The HDMI-I2C controller node can now have various frequency settings.
In other words if due to layout/routing issues the default clock of 100
kHz is to high for stable operation, the clock could be lowered to a
supported frequency for that board.

Supporting peripherals on the other end of the bus. Some monitors
support various sensors or controls over the I2C bus. Think monitor
orientation, back light, brightness, touch etc control. By having a separated
node it becomes possible to configure these devices. Note, that this is
a bad example in the sense that while a monitor (HDMI) supports hot plugging,
I2C device detection does not. This is a potentially future feature,
e.g. a hotplug event re-checks the bus and re-probes I2C drivers on
the bus.

Not working properly: Legacy clocking. For some reason when using an old
devicetree the clocks are not setup properly.

A few future improvements are left TODO:
* Re-probe i2c bus after hot plugging of the HDMI connector.
* I2C clock recovery

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>