input: of_touchscreen: rename touchscreen-size-[xy]
The parameter touchscreen-size-x sits oddly next to
touchscreen-max-pressure for example. Further more when considering
that the actual parameter modified in absinfo is the 'maximum'.

Let us thus rename the touchscreen-max-size-[xy] property in the
devicetree. The new name takes precedence over the earlier
touchscreen-size-[xy] in case both are found.
The former is now deprecated.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>