input: of_touchscreen: add support for minimum sizes
In certain configurations, touchscreens are bigger then the underlying
display. This can happen for various reasons, but an interesting design
choice is to improve the edge detection of touch events. To make sure
we don't have strange offsets we want to pass this information along to
the input framework via the of_touchscreen layer. To this, we need to
also parse the 'minimum' property.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>