input: edt-ft5x06: take into account the number of fingers
It may happen that we miss a touch !down event, for example the
interrupt fires, but we are to late to read the correct data. When this
happens, we may wrongfully set/clear ABS_MT_TRACKING_ID to -1.

The M09 firmware actually tells us how many fingers are being touched.
By using that, we also thus know how many events to focus on and to
clear unused tracking id's. This also has as a boon that we save some
processing time by not evaluating the registers for unused touches.

Note: this should not be used to reduce the number of bytes read, e.g.
read the number of touches register, followed by a read for the actual
registers, as that would result in a non-atomic read, and our data
could be out of sync. E.g. we are told to 4 inputs have been pressed,
but by the time we actually read these registers, 5 inputs are actually
being touched.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>