fbdev: ssd1307fb: Make output a little more consistent
The ssd1307fb driver is a little inconsistent in the debug messages.
Common is to not capitalize the first letter and to not add any periods
etc at the end. Also add a quick shortcut to access the device
structure shortening our error messages.

Finally take this cleanup opportunity to add the missing device.h
header.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>