serial: 8250_dw: Make register access consistent
Almost all register access in the 8250_dw glue is performed on the
byte aligned register address and has a regshift applied in the
serial_in assessor helper. In dw8250_setup_port however we directly
write the register, ignoring the register shift. While it is very
likely that all devices using these offsets are word aligned, it is far
more consistent and readable with the rest of the code base to apply
the register shift and define the registers as their counterparts, byte
aligned.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>