serial: 8250: Only confirm handled interrupts on supported ones
The serial8250_handle_irq function handles many of the possible
interrupts, but none of the newer interrupts. Currently, the function
only notifies of this for the case of the 'no interrupt'. All other
cases are pretended to be handled and the caller has to figure out what
could or is not handled.

Instead, we should check which of the requested interrupts we actually
do handle, and return false otherwise, allowing the caller to take
further measures.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>