serial: 8250: expand debug ability of the irq handler
The 'too much work for irq' error (has) happened often without
understanding whats going on. Over the years several fixes have been
added to address this issue, but even the latest fix is from not so
long ago. To get just a little bit more information from the system
when this happens, also print the IIR register in the debug statement
to help debuggers understand what could possibly go on.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>