serial: 8250: Do not treat the IIR register as a bitfield
In the original 8250 bits 0, 1, 2 and 3 where used together as the
Interrupt Identification Register. Looking at various implementations
and tables it could be argued that bits 1, 2 and 3 where the interrupt
ID and bit 0 was a bit field. This worked for a while, albeit a little
bit messy, but got more confusing as manufactures added more features
to their 8250 cores, such as the designware UART IP, where the entire
nibble was used as an identifier.

The previous premises that bit 0 is special and that we are dealing
with a bit field register is thus no longer true (for certain
implementations).

The IIR register thus always requires a mask to be read on the 8250.
To make this easier we add a helper, as recommended by Andy Shevchenko,
which takes care of the generic 8250 masking. By also allowing an
additional mask, for the peripherals to use more (the reserved) bits,
the helper optionally allows for additional masking.

We currently only do this for the 8250 specific driver, though it is
very likely this applies to others as well.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>