drm: sun4i: improve header readability
By replacing the magic hardcoded bit masks with GENMASK we can improve
readability quite a bit. By also changing the order of shifting and
masking, we can actually use a GENMASK as described in the datasheet
making the masks even more obvious and easier to read.

Furthermore we replace some magic values in the code with defined a
getter in turn use it to improve readability.

This patch does not have any functional changes.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>