drm/sun4i: enable/disable the tmds clock on use
While the tmds clock is configured using set_rate whenever we
use/change it, we never properly 'enable' or 'disable' it throwing off
our ref-counters. The reason it works is because it is an ungated clock
and thus always runs whenever its parent runs.

Add the enable/disable calls to the tmds clock to ensure our
ref-counters play nicely.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>