drm: sun4i: make fifo_transfer more verbose in case of error
The fifo_transfer call currently makes an assumption on the clockrate
of the hdmi-i2c bus. However this assumption is based on the fact that
the clockrate is exactly what we request, e.g. if we set the rate to 100 kHz
the clock runs on exactly 100 kHz. This unfortunately is not exactly
true, as the actual rate depends on the parents and we may end up with
a slower clock. Further more if the clock changes from the 100 kHz now
this assumption may also not hold.

The remedy is easily fixed by asking the clk framework the actual
clockrate, with adding 8 clocks (1 byte) as a safety margin.

Further more, in case of error, it really helps to know that this
timeout has occurred so add a print statement for that.

Finally, the variable was renamed to more accurately tell us what it is,
a byte time in micro seconds, not nano seconds.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>