fbdev: ssd1307fb: Use resolution instead of dimensions
THe virtual resolution and line-length are attributes of the X and Y
resolution, which are determined by the width/height of the display.

Currently it is all directly determined from the width/height
parameter. The width and height never change of the display, but the X,
Y resolution can change when rotating the display.

It thus makes more sense to determine the line length and virtual
parameters from their resolution counterparts and not the display
dimensions.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>