fbdev: of: Add of parsing helper
This patch was inspired by touchscreen_of where a few helper functions
are used to help in parsing various devicetree parameters and thus
offing a standard way for common items.

Here we add support for the first common parameter, 'rotate' which
determines the initial rotation of a screen.

In the future, this can be expanded to also add support for
width/height for example and others as well.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>