fbdev: ssd1307fb: Ensure display is toggled properly
Currently, we 'enable' the display in our last step of init. The
enable mend of the display however should be the very last step of our
probing sequence, e.g. just before saying 'driver loaded successfully',
as this is the time we are 'ready'.

This patch uses the later introduced ssd1307fb_(un)blank function to
achieve this.

Further more, we want to guarantee that the display is in its off state
during initialization, so after reset, we guarantee that the display is
off.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>