fbdev: ssd1307fb: Add hardware accelerated rotation support
The ssd130x series of chips are very flexible, allowing customers to
use any layout to connect the LCD to the PCB. We can use this in
combination with the different scan-out parameters to flip and rotate
the display at will. E.g. A X and Y flip results in a 180 degree
rotation. Changing the scanout from horizontal to vertical rotates by
90 degree's. This thus gives us hardware accelerated rotation support.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>