fbdev: of: Add parameter to not clear screen on probe
There are situations, where the bootloader is relied on to prepare the
display. For example to show a splash screen. Some drivers however
clear the framebuffer during probing and it may be a while before linux
has the chance to show something on the screen. This is experienced by
the user as a blank screen for potentially long seconds.

This patch adds a generic parameter to the framebuffer devicetree
parser to allow the user to explicitly avoid clearing the framebuffer
during load.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>