fbdev: ssd1307fb: Do not always clear the display on boot
Commit 6e376822ee9e06c5 ("fbdev/ssd1307fb: clear screen in probe") in
combination with commit fdde1a8148d81617 ("fbdev: ssd1307fb: Make reset gpio
devicetree property optiona") changed the previous default behavior of
keeping the display active. This for example is a feature that is used
to show a splash screen from u-boot and leave it on the controller
until linux/an application is ready to show its own information.

We thus let the devicetree control this behavior via the
no-clear-on-probe flag, which skips the clearing of the framebuffer and
thus keeps a splash screen on the device.

The default is left to clear the screen (unset in the devicetree) as
the new default does make sense. If u-boot does not initialize the
screen and does not show a splash-screen garbled displays can occur.

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>