ASoC: activating FLL/support of master mode for WM8904 machine driver
Operating the codec in slave mode with a fixed MCLK leads to recognizable distortions
in the sound if the MCLK is not a multiple of the sampling rate. This patch enables
the operation of the codec in master mode and activates the FLL (frequency locked loop)
in order to eliminate the distortions.

Signed-off-by: Michael Schanz <michael.schanz@congatec.com>
Signed-off-by: Alex de Cabo <alejandro.de-cabo-garcia@congatec.com>