rtc-m41t80: Clear oscillator failure bit on probe
The oscillator failure bit of the flags register defaults to 1 when the
m41t80 RTC device is powered up. If we don't clear it as the manual says
we should then the bit remains set permanently and since commit
05a7f27a ("rtc: m41t80: handle oscillator failure bit") we spam the
kernel console with an error message on every attempt to read the RTC,
for example:

  Welcome to Buildroot
  mips64r6-n32-le-hf-soak login: [   15.534459] rtc-m41t80 0-0068:
  Oscillator failure, data is invalid.
  [   16.526308] rtc-m41t80 0-0068: Oscillator failure, data is invalid.
  [   19.534342] rtc-m41t80 0-0068: Oscillator failure, data is invalid.
  [   20.526327] rtc-m41t80 0-0068: Oscillator failure, data is invalid.

Fix this by clearing the OF flag during probe, as the m41t80 manual
says we ought to.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Signed-off-by: Georg <georg.hartinger@congatec.com>