imx8qm-cgtsmx8: bugfix for imx thermal driver
This driver reads the temperature from the devices and
also determins the trend. The driver does only
support passive and critical trips and handles active
trips with eg an assigned pwm-fan incorrect.

With disabled trend function a default thermal-zone
trend function works as excpected even on active trips.
With this, the pwm-fan regulation with different pwm
levels on different temperatures works with
step-wise thermal-governor.

Signed-off-by: Georg <georg.hartinger@congatec.com>